speed_transformation

+++ speed_transformation/src/lib.rs
@@ -1,5 +1,5 @@
-pub fn km_per_hour_to_meters_per_second(km_h: u64) -> u64 {
-    (km_h / 60 ) * 1000
+pub fn km_per_hour_to_meters_per_second(km_h: f64) -> f64 {
+    (km_h / 60 as f64) * 1000 as f64
 }
 
 #[cfg(test)]
@@ -8,7 +8,7 @@ mod tests {
 
     #[test]
     fn it_works() {
-        let result = km_per_hour_to_meters_per_second(60);
-        assert_eq!(result, 1000);
+        let result = km_per_hour_to_meters_per_second(60 as f64);
+        assert_eq!(result, 1000 as f64);
     }
 }
