+++ speed_transformation/src/main.rs
@@ -3,6 +3,6 @@ use speed_transformation::*;
 fn main() {
     println!(
         "60 Km/H equals: {} meters in minute",
-        km_per_hour_to_meters_per_second(60)
+        km_per_hour_to_meters_per_second(60 as f64)
     );
 }
